src_refactored/requirements.txt
hydra-core>=1.3.0
omegaconf>=2.3.0
openai>=1.0.0
loguru>=0.7.0
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
vector-quantize-pytorch>=1.0.0 